requirements.txt
azure-functions
numpy
pandas
pyodbc
collections-extended
spacy>=2.2.0,<3.0.0
https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-2.2.5/de_core_news_sm-2.2.5.tar.gz#egg=de_core_news_sm
